cloudrun_functions/news_daily_report/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
duckdb==1.1.0
pandas
openai